+++ src/Entity/FireBallEntity.java
@@ -39,12 +39,12 @@ public class FireBallEntity extends Entity{
             Composite c = g.getComposite();
             // g.setComposite(new Additive());
             g.setColor(new Color(255,r.nextInt(255),0,r.nextInt(255)));
-                g.fillArc((X-6)-viewX, (Y-6)-viewY, 12,12, 0, 360); 
+                g.fillArc((int)(X-6)-viewX, (int)(Y-6)-viewY, 12,12, 0, 360); 
             for (int i = 0; i < 4; i++)
             {
                 int e1 = 6-r.nextInt(12), e2 = 6-r.nextInt(12);
                 g.setColor(new Color(255,r.nextInt(255),0,r.nextInt(255)));
-                g.fillArc((X+e1)-viewX, (Y+e2)-viewY, e1, e2, 0, 360); 
+                g.fillArc((int)(X+e1)-viewX, (int)(Y+e2)-viewY, e1, e2, 0, 360); 
             }
             g.setComposite(c);
         }
@@ -74,24 +74,24 @@ public void onServerUpdate(Server lol)
         if (lol.earth.checkCollision(X, Y))
         {
            radius*=4;
-           lol.earth.ground.ClearCircle(X, Y, radius);
-           lol.sendMessage(Server.DIG, ByteBuffer.allocate(40).putInt(X).putInt(Y).putInt(radius));
+           lol.earth.ground.ClearCircle((int)X, (int)Y, radius);
+           lol.sendMessage(Server.DIG, ByteBuffer.allocate(40).putInt((int)X).putInt((int)Y).putInt(radius));
            alive = false;
         }
         if (lol.earth.inBounds(X, Y)&&collided(lol.earth))//lol.earth.ground.cellData[X][Y]==World.WATER
         {
             alive = false;
-            lol.sendMessage(Server.STEAM, ByteBuffer.allocate(40).putInt(X).putInt(Y).putInt(this.MYID));
+            lol.sendMessage(Server.STEAM, ByteBuffer.allocate(40).putInt((int)X).putInt((int)Y).putInt(this.MYID));
         }
 }
     @Override
     public void cerealize(ByteBuffer out) {
         try {
             Server.putString(out,  this.getClass().getName());
-            out.putInt(X);
-            out.putInt(Y);
-            out.putInt(xspeed);
-            out.putInt(yspeed);
+            out.putInt((int)X);
+            out.putInt((int)Y);
+            out.putInt((int)xspeed);
+            out.putInt((int)yspeed);
             out.putInt(maker);
         } catch (Exception ex) {
             Logger.getLogger(ExplosionEntity.class.getName()).log(Level.SEVERE, null, ex);
