+++ src/Entity/FireJumpEntity.java
@@ -40,12 +40,12 @@ public class FireJumpEntity extends Entity{
             Composite c = g.getComposite();
             // g.setComposite(new Additive());
             g.setColor(new Color(255,r.nextInt(255),0,r.nextInt(255)));
-                g.fillArc((X-8)-viewX, (Y-8)-viewY, 16,16, 0, 360); 
+                g.fillArc((int)(X-8)-viewX, (int)(Y-8)-viewY, 16,16, 0, 360); 
             for (int i = 0; i < 6; i++)
             {
                 int e1 = 6-r.nextInt(12), e2 = 6-r.nextInt(12);
                 g.setColor(new Color(255,r.nextInt(255),0,r.nextInt(255)));
-                g.fillArc((X+e1)-viewX, (Y+e2)-viewY, e1, e2, 0, 360); 
+                g.fillArc((int)(X+e1)-viewX, (int)(Y+e2)-viewY, e1, e2, 0, 360); 
             }
             g.setComposite(c);
         }
@@ -78,10 +78,10 @@ public void onServerUpdate(Server lol)
     public void cerealize(ByteBuffer out) {
         try {
             Server.putString(out,  this.getClass().getName());
-            out.putInt(X);
-            out.putInt(Y);
-            out.putInt(xspeed);
-            out.putInt(yspeed);
+            out.putInt((int)X);
+            out.putInt((int)Y);
+            out.putInt((int)xspeed);
+            out.putInt((int)yspeed);
             out.putInt(maker);
         } catch (Exception ex) {
             Logger.getLogger(ExplosionEntity.class.getName()).log(Level.SEVERE, null, ex);
