+++ src/Entity/GustEntity.java
@@ -37,14 +37,14 @@ public class GustEntity extends Entity{
         {
             G.setColor(Color.lightGray);
             int deg = r.nextInt(360);
-            G.fillArc((X-1)-viewX, (Y-1)-viewY, 2, 2, deg, 15);
+            G.fillArc((int)(X-1)-viewX, (int)(Y-1)-viewY, 2, 2, deg, 15);
             deg = r.nextInt(360);
-            G.fillArc((X-2)-viewX, (Y-2)-viewY, 4, 4, deg, 15);
+            G.fillArc((int)(X-2)-viewX, (int)(Y-2)-viewY, 4, 4, deg, 15);
             G.setColor(Color.white);
             deg = r.nextInt(360);
-            G.fillArc((X-3)-viewX, (Y-3)-viewY, 6, 6, deg, 15);
+            G.fillArc((int)(X-3)-viewX, (int)(Y-3)-viewY, 6, 6, deg, 15);
             deg = r.nextInt(360);
-            G.fillArc((X-4)-viewX, (Y-4)-viewY, 8, 8, deg, 15);
+            G.fillArc((int)(X-4)-viewX, (int)(Y-4)-viewY, 8, 8, deg, 15);
         }
     }
 
@@ -57,7 +57,7 @@ public class GustEntity extends Entity{
        }
        for (Player p:apples.playerList)
        {
-           if (maker!=p.ID&&p.checkCollision(X, Y))
+           if (maker!=p.ID&&p.checkCollision((int)X, (int)Y))
            {
                alive = false;
            }
@@ -72,10 +72,10 @@ public class GustEntity extends Entity{
     public void cerealize(ByteBuffer out) {
         try {
             Server.putString(out,  this.getClass().getName());
-            out.putInt(X);
-            out.putInt(Y);
-            out.putInt(xspeed);
-            out.putInt(yspeed);
+            out.putInt((int)X);
+            out.putInt((int)Y);
+            out.putInt((int)xspeed);
+            out.putInt((int)yspeed);
             out.putInt(maker);
         } catch (Exception ex) {
             Logger.getLogger(ExplosionEntity.class.getName()).log(Level.SEVERE, null, ex);
@@ -93,8 +93,8 @@ public class GustEntity extends Entity{
 {
     if (collided(lol.earth))
        {
-           lol.earth.ground.ClearCircle(X, Y, radius);
-           lol.sendMessage(Server.DIG, ByteBuffer.allocate(40).putInt(X).putInt(Y).putInt(radius));
+           lol.earth.ground.ClearCircle((int)X, (int)Y, radius);
+           lol.sendMessage(Server.DIG, ByteBuffer.allocate(40).putInt((int)X).putInt((int)Y).putInt(radius));
            alive = false;
        }
 }
